mecab/requirements.txt
apache-beam[gcp]
mecab-python3
